requirements (2).txt
Flask==2.2.3
gunicorn
werkzeug==2.2.3
google-cloud-datastore
